requirements.txt
Flask
google-generativeai
twilio
python-dotenv